requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple

annotated-doc==0.0.3
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via starlette
argon2-cffi==23.1.0
    # via pwdlib
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
asyncpg==0.30.0
    # via -r requirements.in
bcrypt==4.3.0
    # via pwdlib
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
click==8.3.0
    # via uvicorn
colorama==0.4.6
    # via click
cryptography==46.0.3
    # via pyjwt
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via fastapi-users
fastapi==0.120.1
    # via
    #   -r requirements.in
    #   fastapi-users
fastapi-users==15.0.1
    # via -r requirements.in
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   email-validator
makefun==1.16.0
    # via fastapi-users
numpy==2.3.4
    # via pandas
pandas==2.3.3
    # via -r requirements.in
pwdlib[argon2,bcrypt]==0.2.1
    # via fastapi-users
pycparser==2.23
    # via cffi
pydantic==2.12.3
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.41.4
    # via pydantic
pyjwt[crypto]==2.10.1
    # via fastapi-users
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   pandas
python-dotenv==1.2.1
    # via -r requirements.in
python-multipart==0.0.20
    # via fastapi-users
pytz==2025.2
    # via pandas
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy[asyncio]==2.0.36
    # via
    #   -r requirements.in
    #   sqlalchemy
starlette==0.49.1
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
uvicorn==0.38.0
    # via -r requirements.in
